--- a/src/liballoc_system/lib.rs
+++ b/src/liballoc_system/lib.rs
@@ -48,25 +48,9 @@ use core::ptr::NonNull;
 /// This is based on `malloc` on Unix platforms and `HeapAlloc` on Windows,
 /// plus related functions.
 ///
-/// This type can be used in a `static` item
-/// with the `#[global_allocator]` attribute
-/// to force the global allocator to be the system’s one.
-/// (The default is jemalloc for executables, on some platforms.)
-///
-/// ```rust
-/// use std::alloc::System;
-///
-/// #[global_allocator]
-/// static A: System = System;
-///
-/// fn main() {
-///     let a = Box::new(4); // Allocates from the system allocator.
-///     println!("{}", a);
-/// }
-/// ```
-///
-/// It can also be used directly to allocate memory
-/// independently of the standard library’s global allocator.
+/// This type can be used directly to allocate memory
+/// independently of the standard library’s global allocator,
+/// which may have been changed with the `#[global_allocator]` attribute.
 #[stable(feature = "alloc_system_type", since = "1.28.0")]
 pub struct System;