--- a/src/librustc_target/spec/aarch64_unknown_freebsd.rs
+++ b/src/librustc_target/spec/aarch64_unknown_freebsd.rs
@@ -15,7 +15,6 @@ pub fn target() -> TargetResult {
     base.max_atomic_width = Some(128);
 
     // see #36994
-    base.exe_allocation_crate = None;
 
     Ok(Target {
         llvm_target: "aarch64-unknown-freebsd".to_string(),