Fixes https://github.com/rust-lang/rust/issues/36963

**Do not merge** yet. This PR by itself is likely to regress rustc performance. The purpose of opening it now is to measure by how much. We’ll likely want to figure out https://github.com/rust-lang/rust/issues/51038 and land them around the same time.