@SimonSapin 2018-07-03
@bors try

@Mark-Simulacrum, could you run this on perf.rlo?