@cuviper 2018-07-03
Please specifically note the versions of the system allocator (presumably glibc) when reporting performance.  If possible, it would be nice to have results for both current and older glibc, say Fedora 28 versus RHEL/CentOS 7.  If there's a way I can help generate results, let me know -- I don't have hardware to dedicate to this, but I can do manual runs for this evaluation.