@SimonSapin 2018-07-03
Indeed, results on perf.rust-lang.org are presumably collected on Linux with a single version of glibc. Measuring impact on macOS would be nice too.