@Mark-Simulacrum 2018-07-03
Perf results, once ready, will have been run with Ubuntu GLIBC 2.23-0ubuntu10.