@cuviper 2018-07-03
OK.  The most notable change for glibc came in 2.26:
https://sourceware.org/ml/libc-alpha/2017-01/msg00524.html

For Ubuntu, that would be in artful (17.10) and later.