@SimonSapin 2018-07-03
In this case, getting both sets of results would be interesting.