@SimonSapin 2018-07-04
So I tried running the benchmarks on a mac and got this:

```
thread 'main' panicked at 'assertion failed: has_perf', collector/src/bin/rustc-perf-collector/execute.rs:313:9
```

I assume `has_perf` refers to Linux’s perf tool. Is there a mode for the data collector to get the subset of available measures, possibly only wall clock time?

I also ran it on an Ubuntu 18.04 machine with the `libc6` package at version 2.27-3ubuntu1 and an i7 CPU with 4 cores / 8 threads at 4.6 GHz. http://localhost:2346/compare.html?start=64f8ae08fd5294718466a61f74f6de0135b88035&end=91eac4dd1e41932c605075305758d84406abfaf4 does show results, but there’s so much green (where I would have expected at least some red) that I wonder if I’m comparing data sets measured on different machines.

```
compare	2018-07-03 (64f8ae08)	2000-01-01 (91eac4dd)	% change
 
tokio-webpush-simple-opt
	avg: -22.2%	min: -65.7%	max: -0.4%
unify-linearly-opt
	avg: -16.1%	min: -26.7%	max: -12.0%
helloworld-opt
	avg: -17.2%	min: -22.2%	max: -15.1%
helloworld-debug
	avg: -16.3%	min: -17.5%	max: -15.5%
unify-linearly-debug
	avg: -13.7%	min: -16.3%	max: -12.1%
```

<details>

```
issue-46449-opt
	avg: -5.1%	min: -9.3%	max: -0.8%
coercions-opt
	avg: -5.3%?	min: -8.9%?	max: -3.2%?
unused-warnings-opt
	avg: -6.3%	min: -6.9%	max: -5.3%
unused-warnings-debug
	avg: -5.9%	min: -6.9%	max: -5.2%
coercions-debug
	avg: -4.6%?	min: -6.4%?	max: -3.6%?
html5ever-opt
	avg: -3.3%	min: -5.4%	max: -1.5%
html5ever-debug
	avg: -3.5%	min: -5.3%	max: -1.6%
html5ever-check
	avg: -3.7%	min: -5.2%	max: -1.7%
deep-vector-debug
	avg: -2.9%	min: -4.4%	max: -2.1%
deep-vector-opt
	avg: -2.3%	min: -4.4%	max: -1.4%
tuple-stress-opt
	avg: -2.3%	min: -3.8%	max: -0.0%
script-servo-debug
	avg: -1.6%	min: -3.7%	max: -0.7%
coercions-check
	avg: -2.6%?	min: -3.7%?	max: -1.7%?
tuple-stress-check
	avg: -1.9%	min: -3.3%	max: -0.0%
style-servo-debug
	avg: -2.1%?	min: -3.2%?	max: -0.9%?
tokio-webpush-simple-debug
	avg: -1.8%	min: -3.1%	max: -1.2%
tuple-stress-debug
	avg: -2.0%	min: -3.1%	max: -0.0%
sentry-cli-opt
	avg: -1.2%	min: -3.0%	max: -0.4%
inflate-opt
	avg: -1.9%	min: -2.9%	max: -1.2%
encoding-debug
	avg: -1.6%	min: -2.7%	max: -0.7%
ripgrep-opt
	avg: -1.3%	min: -2.7%	max: -0.4%
unused-warnings-check
	avg: -2.3%	min: -2.7%	max: -1.8%
style-servo-opt
	avg: -1.0%?	min: -2.7%?	max: -0.4%?
syn-debug
	avg: -1.7%	min: -2.6%	max: -1.0%
issue-46449-debug
	avg: -1.8%	min: -2.5%	max: -1.5%
ripgrep-debug
	avg: -1.5%	min: -2.5%	max: -0.6%
regression-31157-opt
	avg: -1.2%	min: -2.5%	max: -0.1%
regex-debug
	avg: -1.4%	min: -2.4%	max: -0.3%
syn-opt
	avg: -0.7%	min: -2.3%	max: 1.9%
deep-vector-check
	avg: -1.9%	min: -2.3%	max: -1.6%
regex-opt
	avg: -1.1%	min: -2.2%	max: -0.3%
sentry-cli-check
	avg: -1.9%	min: -2.1%	max: -1.5%
inflate-debug
	avg: -0.1%	min: -2.1%	max: 0.6%
encoding-check
	avg: -1.7%	min: -2.0%	max: -1.6%
regression-31157-debug
	avg: -1.3%	min: -2.0%	max: -0.7%
encoding-opt
	avg: -1.0%	min: -2.0%	max: -0.2%
ripgrep-check
	avg: -1.6%	min: -2.0%	max: -1.3%
futures-debug
	avg: -1.3%	min: -1.9%	max: -0.9%
serde-debug
	avg: -1.2%	min: -1.9%	max: -0.9%
deeply-nested-debug
	avg: -1.1%	min: -1.9%	max: -0.8%
issue-46449-check
	avg: -1.6%	min: -1.9%	max: -1.4%
piston-image-debug
	avg: -1.1%	min: -1.9%	max: -0.3%
serde-opt
	avg: -1.2%	min: -1.8%	max: -0.8%
regression-31157-check
	avg: -1.3%	min: -1.8%	max: -0.9%
sentry-cli-debug
	avg: -1.0%	min: -1.8%	max: -0.4%
regex-check
	avg: -1.7%	min: -1.8%	max: -1.5%
syn-check
	avg: -1.6%	min: -1.8%	max: -1.4%
deeply-nested-check
	avg: -1.2%	min: -1.7%	max: -0.9%
futures-opt
	avg: -1.2%	min: -1.7%	max: -0.8%
tokio-webpush-simple-check
	avg: -1.4%	min: -1.7%	max: -1.3%
cargo-opt
	avg: -0.7%	min: -1.7%	max: -0.4%
deeply-nested-opt
	avg: -0.6%	min: -1.6%	max: -0.1%
cargo-check
	avg: -1.1%	min: -1.6%	max: -0.8%
webrender-check
	avg: -1.2%	min: -1.6%	max: -1.1%
webrender-debug
	avg: -1.0%	min: -1.5%	max: -0.4%
piston-image-opt
	avg: -0.9%	min: -1.5%	max: -0.3%
style-servo-check
	avg: -1.2%?	min: -1.5%?	max: -0.9%?
clap-rs-opt
	avg: -0.8%	min: -1.4%	max: -0.5%
futures-check
	avg: -1.1%	min: -1.4%	max: -0.8%
piston-image-check
	avg: -1.3%	min: -1.4%	max: -1.1%
clap-rs-check
	avg: -0.9%	min: -1.4%	max: -0.5%
unify-linearly-check
	avg: -1.0%	min: -1.4%	max: -0.6%
cargo-debug
	avg: -0.8%	min: -1.4%	max: -0.5%
webrender-opt
	avg: -1.0%	min: -1.3%	max: -0.7%
helloworld-check
	avg: -1.0%	min: -1.3%	max: -0.6%
inflate-check
	avg: 0.7%	min: -0.9%	max: 1.3%
script-servo-check
	avg: -0.5%	min: -1.2%	max: -0.0%
serde-check
	avg: -0.8%	min: -1.0%	max: -0.5%
clap-rs-debug
	avg: -0.6%	min: -0.8%	max: -0.4%
script-servo-opt
	avg: -0.5%	min: -0.7%	max: -0.2%
```

</details>