@SimonSapin 2018-07-07
The results on the infra look similar to those on my machine… and I still don’t know what to make of them. CC @rust-lang/wg-compiler-performance, @glandium, @nnethercote 

* https://perf.rust-lang.org/compare.html?start=64f8ae08fd5294718466a61f74f6de0135b88035&end=91eac4dd1e41932c605075305758d84406abfaf4
* https://perf.rust-lang.org/compare.html?start=64f8ae08fd5294718466a61f74f6de0135b88035&end=91eac4dd1e41932c605075305758d84406abfaf4&stat=wall-time