@SimonSapin 2018-07-07
TL;DR of the numbers: depending on what is being compiled in what mode, compilation wall clock time difference due do disabling jemalloc is between -42.5% (time nearly halved!) and +10.6%