@alexcrichton 2018-07-08
@SimonSapin oh the perf links you gisted actually were for a [pretty wide range of commits](https://github.com/rust-lang/rust/compare/64f8ae08fd5294718466a61f74f6de0135b88035...91eac4dd1e41932c605075305758d84406abfaf4), but [this comparison URL](https://perf.rust-lang.org/compare.html?start=860d169474acabdc53b9a698f8ce02eba7e0daeb&end=91eac4dd1e41932c605075305758d84406abfaf4&stat=wall-time) is just for the jemalloc changes. 

From that URL some conclusions are:

* Tiny programs tend to compile faster (0.17s -> 0.15s) ish
* Big programs compile a little slower (187s -> 211s for script servo debug)

Overall I think the losses here outweigh the gains (as the gains are all very small and only on small programs) and we probably want to stick with jemalloc on Linux for now. It'd be interesting though to compare with the glibc 2.26 allocator as well though!