@nnethercote 2018-07-09
I agree with @alexcrichton. The ones that improved were toy or stress programs; all the more interesting benchmarks regressed :(