@michaelwoerister 2018-07-09
I agree with @alexcrichton and @nnethercote, we should not merge this as is. Thanks a lot for measuring though, @SimonSapin!