@mati865 2018-07-09
For redis glibc 2.26 [can make huge difference](https://www.phoronix.com/scan.php?page=news_item&px=Glibc-2.26-Redis-Test)

It should be revaluated again once perf.rust-lang.org is updated to something more recent.