@michaelwoerister 2018-07-09
> For redis glibc 2.26 can make huge difference

Those numbers look great but wouldn't that mean that people on older systems would still experience a noticeable regression?